new dataset and train script + adding input and output path to slurm template
