+++ app/controllers/slurm/slurm_template.srm
@@ -16,4 +16,4 @@ JOBNAME=$SLURM_JOB_NAME
 echo "JOBNAME: " $JOBNAME
 
 # TODO: check if container.sif exists
-singularity run -B {folder} {folder}/{image_name}.sif python3 {folder}/{train_script_name} {folder}/{dataset_name}
+singularity run -B {folder} {folder}/{image_name}.sif python3 {folder}/{train_script_name} input_path={folder}/{dataset_name} output_path={folder}
