--- test/submiter_example.json
@@ -1,21 +1,0 @@
-{
-  "dataset_name": "wine_classification",
-  "model_params": {
-    "n_estimators": 2,
-    "random_state": 42
-  },
-  "runner_location": "atena02",
-  "model_tracking": true,
-  "tracking_uri": "experiments",
-  "experiment_name": "atena_test",
-  "train_script_name": "train.py",
-  "clusters": {
-    "atena02": {
-      "infra_config": {
-        "instance_type": "cpu",
-        "image_name": "sklearn_sample_latest",
-        "account": "twinscie"
-      }
-    }
-  }
-}
